% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data_table_to_vw}
\alias{data_table_to_vw}
\title{Format a data.table in vowpal wabbit format}
\usage{
data_table_to_vw(x, target_var = NULL, weight_var = NULL,
  namespaces = NULL, out_digits = 8, verbose = TRUE)
}
\arguments{
\item{x}{a data.table (or object coerceable to data.table)}

\item{target_var}{(optional) the y variable for supervised vowpal wabbit models}

\item{weight_var}{(optional) the weights variable for supervised vowpal wabbit models}

\item{namespaces}{a list of namespaces and the variables in them.  By default this function builds 1 namespace}

\item{out_digits}{How many digests to round numbers to before writing}

\item{verbose}{Print a log as the function runs}
}
\value{
A character vector
}
\description{
This function formats a data.table in vowpal wabbit format (with namespaces).
}

