\name{oneRowDataset}
\alias{oneRowDataset}
\title{Create a single row dataset with all possible VW variables}
\usage{
  oneRowDataset(y, X, namespaces)
}
\arguments{
  \item{y}{}

  \item{X}{}

  \item{namespaces}{}
}
\value{
  a single row data.frame
}
\description{
  This function writes a single-row VW dataset, but
  includes all possible dumy variable levels.  It was
  intended for reversing VW hashes, but is NOW DEPRECIATED!
  It will be removed in future versions of this package
}

