\name{Vw-package}
\alias{Vw-package}
\alias{Vw}
\docType{package}
\title{
\packageTitle{Vw}
}
\description{
\packageDescription{Vw}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Vw}
\packageIndices{Vw}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Vw}

Maintainer: \packageMaintainer{Vw}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
