% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vwHelp}
\alias{vwHelp}
\title{Display the vowpal wabbit help functions}
\usage{
vwHelp(vw_path = getOption("vw_path"))
}
\arguments{
\item{vw_path}{the path to the vw executible}
}
\description{
This function prints the help for the VW executible
}
\examples{
vwHelp()
}

