\name{logit}
\alias{logit}
\title{
Logit transform
}
\description{
Logit transform: log(x/(1-x))
}
\usage{
logit(x)
}

\arguments{
  \item{x}{numeric}
}

\value{
logit transform of x
}

\references{
Manas Geeta Arun, Aidan Angus-Henry, Darren J. Obbard & Jarrod D. Hadfield (2015) "Estimating the additive genetic variance for relative fitness from changes in allele frequency".
}
\author{
Manas Geeta Arun and Jarrod Hadfield
}
