\name{Vw-package}
\alias{Vw-package}
\alias{Vw}
\docType{package}
\title{
\packageTitle{Vw}
}
\description{
\packageDescription{Vw}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Vw}
\packageIndices{Vw}

}
\author{
\packageAuthor{Vw}

Maintainer: \packageMaintainer{Vw}
}
\references{
Manas Geeta Arun, Aidan Angus-Henry, Darren J. Obbard & Jarrod D. Hadfield (2015) "Estimating the additive genetic variance for relative fitness from changes in allele frequency".
}
\keyword{package}
